изменено:      src/Controller/CommentController.php 	удалено:       src/Migrations/Version20200319102546.php 	удалено:       src/Migrations/Version20200325184213.php 	удалено:       src/Migrations/Version20200327162359.php 	удалено:       src/Migrations/Version20200329173124.php 	удалено:       src/Migrations/Version20200330104752.php 	удалено:       src/Migrations/Version20200331113424.php 	удалено:       src/Migrations/Version20200331122427.php 	удалено:       src/Migrations/Version20200331123027.php 	удалено:       src/Migrations/Version20200403134002.php 	удалено:       src/Migrations/Version20200403134723.php 	удалено:       src/Migrations/Version20200403135236.php

+++ src/Controller/CommentController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Entity\User ;
 use App\Entity\Projects ;
 use App\Entity\Ticket;
 use App\Entity\Comment;
@@ -20,26 +21,35 @@ class CommentController extends AbstractController
       */
      public function newcomment ( Request $request, $id ) : Response
      {
-	 $tick = $this -> getDoctrine ($id)
+         $users = $this->getDoctrine()
+            -> getRepository(User::class)
+	    -> findAll();
+
+	 $tick = $this -> getDoctrine ($id)		     
             -> getManager()
             -> getRepository ( Ticket :: class )
             -> find ( $id );
 
-//         $ticketId = $request->query->get('id');
 	 $comment = new Comment();
-	 $comment -> setTicketId($id);
-	 $comment -> setUserId(1);
+	 $creator = $this->getUser()->getId();
+//	 $comment -> setTicketId($id);
+//         $comment -> setUserId(1);
          $form = $this->createForm(CommentType::class, $comment);
-         $form->handleRequest($request);
+	 $form->handleRequest($request);
+
          if ($form->isSubmitted() && $form->isValid()) {
              $entityManager = $this->getDoctrine()->getManager();
-//             $ticket = $entityManager->getRepository(Ticket::class)->find($ticketId);
+             
+             $user = $entityManager -> getRepository(User::class)->find($creator);
+             $ticket = $entityManager -> getRepository(Ticket::class)->find($id);
 
-//             $comment->setProject($tick);
+	     $comment -> setTicket($ticket);
+             $comment -> setCreator($user);
 
-//             $entityManager->persist($tick);
-             $entityManager->persist($comment);
-             $entityManager->flush();
+	     $entityManager -> persist($ticket);
+	     $entityManager -> persist($user);
+             $entityManager -> persist($comment);
+             $entityManager -> flush();
 
              return $this->redirectToRoute('comment', ['id' => $id]);
          }
@@ -56,6 +66,7 @@ class CommentController extends AbstractController
       */
      public function comments( Request $request, $id) : Response
      {
+         $user = $this->getUser()->getId();
 	 $tick = $this -> getDoctrine ($id)
             -> getManager()
             -> getRepository ( Ticket :: class )
@@ -64,28 +75,46 @@ class CommentController extends AbstractController
         	 
          $comments = $this->getDoctrine()
              -> getRepository(Comment::class)
-             -> findByTicketld($id);
-
+             -> find($id);
+ //        $user = $comments -> 
 	 return $this->render('Ticket/comment.html.twig', [
 		     'tick' => $tick,
 		     'comment' => $comments,
+		     'user' => $user,
          ]);
      }
 
      /**
-      * @Route("/delete_com/{id}", name="delete_com")
+      * @Route("/{tick_id}/delete_com/{id}", name="delete_com")
       */
-     public function delete_com ( Request $request, $id ) : Response
+     public function delete_com ($tick_id, Request $request, $id ) : Response
      {
-         $ticketId = $request->query->get('id');
+	  $tick = $this -> getDoctrine ($tick_id)
+            -> getManager()
+            -> getRepository ( Ticket :: class )
+            -> find ( $tick_id );
+	  //         $ticketId = $request->query->get('id');
+	 $comments = $this->getDoctrine()
+             -> getRepository(Comment::class)
+             -> find($tick_id);
+
          $comment = $this -> getDoctrine ($id)
             -> getRepository ( Comment :: class )
             -> find ( $id );
-         $em = $this->getDoctrine()->getManager();
-         $em -> remove($comment);
-         $em -> flush();
+	  
+	 if($comment){ 
+	    $em = $this->getDoctrine()->getManager();
+            $em -> remove($comment);
+            $em -> flush();
+         
+            return $this->redirectToRoute('comment', ['id' => $tick_id]);
+         }
 
-         return $this->redirectToRoute('comment', ['id' => $ticketId]);
+	  return $this->render('Ticket/comment.html.twig', [
+                     'tick' => $tick,
+                     'comment' => $comments,
+         ]);
+ 
       }
 }
